Move operation type into _types/operation.py
