+++ tamr_client/dataset/unified.py
@@ -4,9 +4,8 @@ See https://docs.tamr.com/reference/dataset-models
 from copy import deepcopy
 
 from tamr_client import operation, response
-from tamr_client._types import JsonDict, UnifiedDataset, URL
+from tamr_client._types import JsonDict, Operation, UnifiedDataset, URL
 from tamr_client.instance import Instance
-from tamr_client.operation import Operation
 from tamr_client.project import Project
 from tamr_client.session import Session
